requirements.txt
google-generativeai
faiss-cpu
numpy
pandas
python-dotenv
coloredlogs
tqdm